fix(ops): auto-ingest negative Google reviews, then nudge
Root cause of 'bad reviews not sent': negatives were only ingested into
ReviewReplyDraft when a human opened the reviews board (sync=1) — no cron did it.
The daily review crons only auto-reply to positives + snapshot counts. So new
negatives sat un-ingested and the nudge had nothing to send.

Extracted the board's sync into lib/reviews/sync-negatives (syncNegativeReviewDrafts);
the review-nudge cron now ingests new GBP negatives first, then nudges on-shift team
+ managers. The board route reuses the same function (no behaviour change).
